Cleaned up helper functions, began packaging them up
