+++ case.py
@@ -27,8 +27,7 @@ class case:
         self.timeseries = pd.DataFrame(index = timeseries_index, columns = timeseries_columns)
 
         self.time_vector = np.linspace(0, forecast_duration, forecast_duration + 1) * self.settings['days_in_month']
-        self.timeseries['days_start'] = self.time_vector[:-1]
-        self.timeseries['days_end'] = self.time_vector[1:]
+        self.timeseries['days_start'], self.timeseries['days_end'] = dca.vector_to_endpoints(self.time_vector)
     
     def gas_forecast(self, forecast_type, qi = None, qf = None, De = None, Dte = None, b = None):
         
@@ -39,6 +38,7 @@ class case:
 
         dispatch_map = {
             'exponential': dca.calc_exponential_forecast,
+            'harmonic': dca.calc_harmonic_forecast
         }
         
         gas_forecast = dispatch_map[forecast_type](self.time_vector, qi, Di)
