+++ decline_curve_analysis/settings.json
@@ -1,0 +1,6 @@
+{
+    "days_in_year": 365.25,
+    "days_in_month": 30.4375,
+    "forecast_duration": 600,
+    "effective_date": "2022-01"
+}
